Use createSlice from RTK for pagination
