--- src/store/pagination/reducer.ts
@@ -1,22 +1,0 @@
-import { createReducer } from '@reduxjs/toolkit';
-import { IPagination } from '../../types/IPagination';
-import { loadFilteredGuitars } from '../guitars/action';
-import { loadTotalPageCount, loadCurrentPage } from './action';
-
-export const initialState: IPagination = {
-  currentPage: 1,
-  totalPageCount: 1,
-};
-
-export const paginationReducer = createReducer(initialState, (builder) => {
-  builder
-    .addCase(loadCurrentPage, (state, action) => {
-      state.currentPage = action.payload;
-    })
-    .addCase(loadTotalPageCount, (state, action) => {
-      state.totalPageCount = action.payload;
-    })
-    .addCase(loadFilteredGuitars, (state, action) => {
-      state.totalPageCount = action.payload.totalPageCount;
-    });
-});
