+++ src/store/store.ts
@@ -1,10 +1,9 @@
 import { configureStore } from '@reduxjs/toolkit';
 import { axiosInstance } from '../api/api';
-// import { filtersReducer } from './filters/reducer';
+// import guitarsReducer from './guitars/slice';
 import { guitarsReducer } from './guitars/reducer';
-// import { pagination } from './pagination/slice';
+import paginationReducer from './pagination/slice';
 import filtersReducer from './filters/slice';
-import { paginationReducer } from './pagination/reducer';
 
 export const store = configureStore({
   reducer: {
